+++ examples/redis_example/redis_store.js
@@ -81,7 +81,10 @@ function redisStore(args) {
         });
     };
 
-    self.del = function(key, cb) {
+    self.del = function(key, options, cb) {
+        if (typeof options === 'function') {
+            cb = options;
+        }
         connect(function(err, conn) {
             if (err) { return cb(err); }
             conn.del(key, handleResponse(conn, cb));